Add headers to KafkaAcknowledgedMessage
When a message is being sent to Kafka the acknowledgement should also contains the headers.
